Create GPU rendering to Default Framebuffer only
Replace Runner::Context with Target.
OutOfPlaceProcessor is now used to render to
OpenGL Default Framebuffer only and can't return the processed image.
ReturningProcessor has old OutOfPlaceProcessor functionality
i.e. copies, processes and returns the image.
Create Display which is associated with a GLFW window and can be a
target for a Sequential-/ConcurrentSink.